rename to NettyConnectionInstrumentationFlag
