+++ instrumentation/netty/netty-4-common/library/src/main/java/io/opentelemetry/instrumentation/netty/v4/common/internal/client/NettyConnectionInstrumentationFlag.java
@@ -9,12 +9,12 @@ package io.opentelemetry.instrumentation.netty.v4.common.internal.client;
  * This class is internal and is hence not for public use. Its APIs are unstable and can change at
  * any time.
  */
-public enum NettyInstrumentationFlag {
+public enum NettyConnectionInstrumentationFlag {
   ENABLED,
   ERROR_ONLY,
   DISABLED;
 
-  public static NettyInstrumentationFlag enabledOrErrorOnly(boolean b) {
+  public static NettyConnectionInstrumentationFlag enabledOrErrorOnly(boolean b) {
     return b ? ENABLED : ERROR_ONLY;
   }
 }
